requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
torchvision==0.15.2
gymnasium==0.29.0
stable-baselines3==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
ta==0.10.2
plotly==5.15.0
tqdm==4.65.0
joblib==1.3.1
psutil==5.9.5
python-dotenv==1.0.0
tensorboard==2.13.0
ray[tune]==2.6.3
optuna==3.2.0 